requirements.txt
Django==4.2.7
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.7.0
mysqlclient==2.2.0
beautifulsoup4==4.12.2
requests==2.31.0